Simple converter to BDF format
Converts all fonts in a file to BDF format.  The files created are named
<face name>-<point size>-<pixel height>-<vert. resolution>.bdf.
Warning: files are overwritten without confirmation!

So far this has only been tested with system fonts and Alex Taylor's fonts.
There are still a number of remaining FIXMEs, but this should already be
usable.

Based on os2font.c.